requirements.txt
conch==8.1.0
numpy==1.21.5
pandas==1.4.4
Pillow==12.0.0
scipy==1.9.1
torch==2.2.0
torchvision==0.24.0
tqdm==4.64.1
